Fix bullet when zoom

+++ include/Bullet.h
@@ -8,7 +8,7 @@ class Window;
 
 class Bullet {
 public:
-    Bullet(Window* window, int originX, int originY, float directionX, float directionY, float speed, int damage, int range);
+    Bullet(Window* window, float originX, float originY, float directionX, float directionY, float speed, int damage, int range);
     bool update(float deltaTime, std::vector<Zombie*> zombies);
     // void draw(SDL_Renderer* renderer, int cameraX, int cameraY, int windowWidth, int windowHeight);
     void draw(int cameraX, int cameraY);
@@ -20,7 +20,7 @@ public:
 
 private:
     Window* window;
-    int originX, originY;
+    float originX, originY;
     float x, y;
     float directionX, directionY;
     float speed;
